Move some files realted to the team and projects to separate directories.
