+++ src/App.js
@@ -3,8 +3,8 @@ import { TopScreen } from './components/TopScreen';
 // import { WhatWeDo } from './components/WhatWeDo';
 import { MovingLines } from './components/MovingLines';
 import { Story } from './components/Story';
-import { OurTeam } from './components/OurTeam';
-import { OurWorks } from './components/OurWorks';
+import { OurTeam } from './components/team/OurTeam';
+import { OurWorks } from './components/works/OurWorks';
 import { Banner } from './components/Banner';
 import { Footer } from './components/Footer';
 import "./scss/main.css";
